inventory-store/requirements.txt
flask
flask_cors
google-cloud-firestore
datetime
uuid
gunicorn
